Modified ResultsView and its controller.
Changed the error handling in the results view so that the user cannot enter negative values for the requested range. 
Changed the drawing logic so that the text doesn't go off screen when the requested values are too high or too low. 
Modified the calculations in the controller to handle exceptions that were thrown when the requested q values were not accessible for the given beamline configuration.